Refactor FileType imports and update type handling
- Updated import paths for `FileType` to reflect new locations and organization.
- Refactored type handling methods to use new `FileType` class for consistency.
- Replaced direct string comparisons with method calls on `FileType` instances to improve type safety and maintainability.
- Consolidated scattered type validation logic into centralized `FileType` methods.
- Introduced getter and setter methods for `type` property in document, note, project, and task data classes for better encapsulation.
