+++ src/models/Data/TopicData.ts
@@ -5,6 +5,7 @@ import IPrjTaskManagement from 'src/interfaces/IPrjTaskManagement';
 import { ITag } from 'src/libs/Tags/interfaces/ITag';
 import { ITags } from 'src/libs/Tags/interfaces/ITags';
 import { Tags } from 'src/libs/Tags/Tags';
+import { FileType } from 'src/types/FileType/FileType';
 import {
     FileSubType,
     Status,
@@ -21,8 +22,25 @@ export default class TopicData
     extends BaseData<TopicData>
     implements IPrjData, IPrjTaskManagement
 {
+    /**
+     * The type of the Topic.
+     */
+    private _type: FileType | null | undefined;
+
+    /**
+     * Sets the type of the Topic.
+     */
     @fieldConfig('Topic')
-    type: 'Topic' | null | undefined;
+    set type(value: unknown) {
+        this._type = new FileType(value);
+    }
+
+    /**
+     * Gets the type of the Topic.
+     */
+    get type(): FileType | null | undefined {
+        return this._type?.valueOf() as FileType | null | undefined;
+    }
 
     @fieldConfig()
     subType: FileSubType | null | undefined;
